Delete assets when the page moves

+++ api/routes/pages/delete.php
@@ -7,15 +7,14 @@ return [
     'method'  => 'DELETE',
     'action'  => function ($path) {
 
-        // try to find the page
-        $page = $this->site()->find($path);
+        if ($page = $this->site()->find($path)) {
+            // delete all assets for the page
+            $assets = new PageAssets($this->app()->root('files'), $page);
+            $assets->delete();
 
-        // delete all assets for the page
-        $assets = new PageAssets($this->app()->root('files'), $page);
-        $assets->delete();
-
-        // delete the page folder
-        return $page->delete();
+            // delete the page folder
+            return $page->delete();
+        }
 
     }
 ];
